requirements.txt
# Core
fastapi==0.121.0
uvicorn[standard]==0.38.0
scikit-learn==1.7.2
openai>=2.6.1
python-dotenv==1.1.1
PyJWT[crypto]>=2.10.1

# Testing
pytest==9.0.2
pytest-cov==7.0.0
httpx==0.28.1

# Mutation testing
mutmut==3.5.0
